requirements.txt
python==3.7.10
pytorch==1.8.1
scipy==1.4.1
numpy==1.19.4
librosa==0.8.1
pyword==0.2.11
joblib==0.14.1





